Move `RsBackgroundTaskQueue` to `org.rust.util`
